+++ packages/feflow-cli/src/core/universal-pkg/repository/git.ts
@@ -49,16 +49,19 @@ export async function getCurrentTag(
   repoPath: string
 ): Promise<string | undefined> {
   const status = spawn.sync('git', ['-C', repoPath, 'status']);
-  const tagFlag = 'HEAD detached at ';
   const head = status?.stdout?.toString().trim().split('\n')[0];
-  if (head.startsWith(tagFlag)) {
-    return head.substring(tagFlag.length);
+  const fields = head.split(' ');
+  if (fields.length > 0) {
+    const v = fields[fields.length - 1];
+    if (versionImpl.check(v)) {
+      return v;
+    }
   }
 }
 
 export function checkoutVersion(repoPath: string, version: string) {
   const command = 'git';
-  spawn.sync(command, ['-C', repoPath, 'pull'], { stdio: 'ignore' });
-  const checkArgs = ['-C', repoPath, 'checkout', version];
+  spawn.sync(command, ['-C', repoPath, 'fetch', '--tags', '-f'], { stdio: 'ignore' });
+  const checkArgs = ['-C', repoPath, 'checkout', '-f', version];
   return spawn.sync(command, checkArgs, { stdio: 'ignore' });
 }