+++ packages/feflow-cli/src/shared/git.ts
@@ -1,14 +1,32 @@
 import spawn from 'cross-spawn';
 
+let isSSH: any;
 async function isSupportSSH(url: string): Promise<any> {
-  const ret = spawn.sync('ssh', ['-vT', url]);
-  const stderr = ret?.stderr?.toString();
-
-  if (/Authentication succeeded/.test(stderr)) {
-    return true;
+  if (isSSH) {
+    return isSSH;
   }
+  try {
+    const res: any = await Promise.race([
+      spawn.sync('ssh', ['-vT', url]),
+      new Promise((resolve: any, reject: any) => {
+        setTimeout(() => {
+          reject(new Error('SSH check timeout'));
+        }, 1000);
+      })
+    ]);
 
-  return false;
+    const stderr = res?.stderr?.toString();
+    if (/Authentication succeeded/.test(stderr)) {
+      isSSH = true;
+    } else {
+      isSSH = false;
+    }
+    return isSSH;
+  } catch (err) {
+    console.log('Git ssh check timeout, use https');
+    isSSH = false;
+    return isSSH;
+  }
 }
 
 function getHostname(url: string): string {
@@ -21,7 +39,8 @@ function getHostname(url: string): string {
   }
 }
 
-export async function transformUrl(url: string): Promise<any> {
+let gitAccount: any;
+export async function transformUrl(url: string, account?: any): Promise<any> {
   const hostname = getHostname(url);
   const isSSH = await isSupportSSH(`git@${hostname}`);
   if (isSSH) {
@@ -31,10 +50,19 @@ export async function transformUrl(url: string): Promise<any> {
       return url;
     }
   } else {
+    let transformedUrl;
     if (/https?/.test(url)) {
-      return url;
+      transformedUrl = url;
     } else {
-      return url.replace(`git@${ hostname }:`, `http://${ hostname }/`);
+      transformedUrl = url.replace(`git@${ hostname }:`, `http://${ hostname }/`);
     }
+    if (account) {
+      gitAccount = account;
+    }
+    if (gitAccount) {
+      const { username, password } = gitAccount;
+      return transformedUrl.replace(/http:\/\//, `http://${username}:${password}@`);
+    }
+    return transformedUrl;
   }
 }