+++ packages/feflow-report/src/constants.ts
@@ -8,6 +8,15 @@ export const HOOK_TYPE_BEFORE = 'before';
  */
 export const HOOK_TYPE_AFTER = 'after';
 
-export const REPORT_URL = 'http://193.112.196.252/api/v1/report/command';
+const BASIC_URL = 'http://api.feflowjs.com';
+
+export const REPORT_URL = `${BASIC_URL}/api/v1/report/command`;
 
 export const REPORT_PROXY = 'http://127.0.0.1:12639';
+
+export const TIMEOUT = 1000;
+
+export const REPORT_STATUS = {
+  START: 0,
+  COMPLETED: 1,
+};