+++ packages/feflow-report/src/index.ts
@@ -1,11 +1,12 @@
 import ApiController from './api';
-import { getUserNameFromGit, getSystemInfoByOS, getProjectByPackage } from './common/utils';
+import { getUserName, getSystemInfoByOS, getProjectByPackage, getProjectByGit } from './common/utils';
 import objectFactory from './common/objectFactory';
-import { HOOK_TYPE_BEFORE, HOOK_TYPE_AFTER } from './constants';
+import { HOOK_TYPE_BEFORE, HOOK_TYPE_AFTER, REPORT_STATUS } from './constants';
 
 interface ReportContext {
   log: any;
   logger: any;
+  args: any;
   pkgConfig: {
     name: string;
   };
@@ -15,6 +16,7 @@ interface ReportContext {
   hook: {
     on(eventName: string, listener: any): void;
   };
+  version: string;
 }
 
 interface ReportBody {
@@ -32,27 +34,22 @@ class Report {
   cmd: string;
   args: object;
   isRecallActivating: boolean;
-  cmdSource: string;
+  commandSource: string;
 
   constructor(feflowContext: ReportContext, cmd?: string, args?: any) {
     this.ctx = feflowContext;
     this.cmd = cmd;
     this.args = args;
-    this.userName = this.getUserName();
+    this.userName = getUserName();
     this.systemInfo = this.getSystemInfo();
     this.project = this.getProject();
     this.loadContextLogger();
-
-    // hook is not supported in feflow 0.16.x
-    if (this.ctx.hook) {
-      this.registerHook();
-    }
   }
   // register before/after hook event
   private registerHook() {
     const { cmd, args } = this;
-    this.ctx.hook.on(HOOK_TYPE_BEFORE, () => {      
-      this.cmdSource = this.ctx.commander?.store[cmd]?.pluginName || '';
+    this.ctx.hook.on(HOOK_TYPE_BEFORE, () => {
+      this.commandSource = this.ctx.commander?.store[cmd]?.pluginName || '';
       this.ctx.log.debug('HOOK_TYPE_BEFORE');
       this.startTime = Date.now();
       this.report(cmd, args);
@@ -70,21 +67,18 @@ class Report {
     this.ctx.log = this.ctx.log ? this.ctx.log : { info: console.log, debug: console.log };
   }
   private getProject() {
-    const { pkgConfig } = this.ctx;
+    const pkgConfig: any = this.ctx.pkgConfig || {};
     let project = '';
 
-    if (pkgConfig) {
+    if (pkgConfig.name) {
       // feflow context
       project = pkgConfig.name;
     } else {
-      // if not, read project name from project's package.json
-      project = getProjectByPackage();
+      // if not, read project name from project's package.json or git
+      project = getProjectByPackage() || getProjectByGit();
     }
 
     return project;
-  }
-  getUserName() {
-    return getUserNameFromGit();
   }
 
   getSystemInfo(): string {
@@ -96,11 +90,13 @@ class Report {
     return objectFactory
       .create()
       .load('command', cmd)
-      .load('cmd_source', this.cmdSource)
+      .load('feflow_version', this.ctx.version)
+      .load('command_source', this.commandSource)
       .load('user_name', this.userName)
       .load('params', args)
       .load('system_info', this.systemInfo)
       .load('project', this.project)
+      .load('status', REPORT_STATUS.START)
       .done();
   }
 
@@ -111,19 +107,27 @@ class Report {
       .load('recall_id', this.reCallId)
       .load('cost_time', this.costTime)
       .load('is_fail', false)
+      .load('status', REPORT_STATUS.COMPLETED)
       .done();
   }
 
   private checkBeforeReport(cmd, args) {
     return !!cmd;
   }
-
+  init(cmd: string) {
+    this.cmd = cmd;
+    this.args = this.ctx.args;
+    // hook is not supported in feflow 0.16.x
+    if (this.ctx.hook) {
+      this.registerHook();
+    }
+  }
   report(cmd, args?) {
     // args check
     if (!this.checkBeforeReport(cmd, args)) return;
     try {
       const reportBody: ReportBody = this.getReportBody(cmd, args);
-      this.ctx.log.debug('reportBody', reportBody);
+      this.ctx.log.debug('reportBody', JSON.stringify(reportBody));
       const report = new ApiController(reportBody, this.ctx.log);
       report.doReport(({ result }) => {
         const { id } = result || {};
@@ -143,7 +147,7 @@ class Report {
     if (!this.reCallId) return;
     try {
       const reCallBody: ReportBody = this.getRecallBody();
-      this.ctx.log.debug('reCallBody', reCallBody);
+      this.ctx.log.debug('reCallBody', JSON.stringify(reCallBody));
       const report = new ApiController(reCallBody, this.ctx.log);
       report.doReport();
     } catch (error) {